Readjusting of Files and Organization.
A lot of what I did was moving around assets and renaming some files and organize everything. If it's Dependent, that means that it's blueprints, levels, and level sequencing. These means the files can be left out and not in a folder and it's critical to the game: GAME CAN'T RUN WITHOUT THESE!!
If it's an independent resource (I have a file for that is ignored by GitHub), these mean that the game can run dependently without them. I'll zip it up every Tuesday and Thursday and send them out in GoFiles for y'all to access them.